Remove welcome as a title (#1798)
A website already welcomes your users. So, you do not actually want to write the word welcome as the page title. Instead, we title the page with what the page is about. Here's more info on this best practice. It's #6:

https://www.nngroup.com/articles/top-10-mistakes-web-design/

I'd probably prefer "Get started with Gitpod" as the title, but I didn't want to remove the cheeky voice you use for docs.